phase10(sql-d): parser-only acceptance for CTEs, views, triggers, windows
Tier-2 / Tier-3 surface coverage. The SQL parser now recognises the
following SQLite-compat forms and the executor surfaces a structured
"parsed-only" error for each, so SQL text round-trips cleanly through
schema-migration tools:

- WITH ... SELECT (CTE), including WITH RECURSIVE.
- CREATE VIEW v AS SELECT ...
- CREATE TRIGGER ... FOR EACH ROW BEGIN ... END.
- Function calls with an OVER() window clause (ROW_NUMBER, RANK, ...).
- Column-level GENERATED ALWAYS AS (...) STORED / VIRTUAL.

datetime.rs satisfies clippy strict-warnings (wrong_self_convention on
to_unix and unnecessary i32 casts in Zeller's congruence).

Tests: 10 new (3 CTE, 1 VIEW, 1 TRIGGER, 3 generated, 2 window).
Workspace baseline: 248 -> 288 tests, all green.

+++ crates/sql/src/datetime.rs
@@ -51,7 +51,11 @@ impl DateTime {
         let doy = doe - (365 * yoe + yoe / 4 - yoe / 100);
         let mp = (5 * doy + 2) / 153;
         let day = (doy - (153 * mp + 2) / 5 + 1) as u32;
-        let month = if mp < 10 { (mp + 3) as u32 } else { (mp - 9) as u32 };
+        let month = if mp < 10 {
+            (mp + 3) as u32
+        } else {
+            (mp - 9) as u32
+        };
         let year = (y + if month <= 2 { 1 } else { 0 }) as i32;
         Self {
             year,
@@ -65,6 +69,7 @@ impl DateTime {
         }
     }
 
+    #[allow(clippy::wrong_self_convention)]
     pub fn to_unix(&self) -> i64 {
         let y = self.year as i64 - if self.month <= 2 { 1 } else { 0 };
         let era = y.div_euclid(400);
@@ -107,11 +112,7 @@ impl DateTime {
     }
 
     pub fn format_datetime(&self) -> String {
-        format!(
-            "{} {}",
-            self.format_date(),
-            self.format_time(),
-        )
+        format!("{} {}", self.format_date(), self.format_time(),)
     }
 }
 
@@ -352,11 +353,7 @@ pub fn strftime(format: &str, dt: &DateTime) -> String {
                 b's' => out.push_str(&dt.to_unix().to_string()),
                 b'w' => out.push_str(&day_of_week(dt).to_string()),
                 b'%' => out.push('%'),
-                b'f' => out.push_str(&format!(
-                    "{:02}.{:03}",
-                    dt.second,
-                    dt.micro / 1000
-                )),
+                b'f' => out.push_str(&format!("{:02}.{:03}", dt.second, dt.micro / 1000)),
                 b'J' => out.push_str(&format!("{:.6}", dt.julian_day())),
                 other => {
                     out.push('%');
@@ -386,13 +383,9 @@ fn day_of_week(dt: &DateTime) -> u32 {
     } else {
         dt.month
     };
-    let y = if dt.month < 3 {
-        dt.year - 1
-    } else {
-        dt.year
-    };
-    let k = (y % 100) as i32;
-    let j = (y / 100) as i32;
+    let y = if dt.month < 3 { dt.year - 1 } else { dt.year };
+    let k = y % 100;
+    let j = y / 100;
     let h = (dt.day as i32 + (13 * (m as i32 + 1)) / 5 + k + k / 4 + j / 4 + 5 * j) % 7;
     ((h + 6) % 7) as u32
 }
